decouple JitDecimalSumState and JitDecimalAvgState
